Move storage related code into one folder
