+++ src/client/datascience/interactive-common/notebookProvider.ts
@@ -10,7 +10,7 @@ import { IWorkspaceService } from '../../common/application/types';
 import { IDisposableRegistry, Resource } from '../../common/types';
 import { noop } from '../../common/utils/misc';
 import { Identifiers } from '../constants';
-import { INotebookStorageProvider } from '../interactive-ipynb/notebookStorageProvider';
+import { INotebookStorageProvider } from '../notebookStorage/notebookStorageProvider';
 import { KernelSpecInterpreter } from '../jupyter/kernels/kernelSelector';
 import {
     ConnectNotebookProviderOptions,
