+++ src/client/datascience/notebookStorage/nativeEditorProvider.ts
@@ -42,7 +42,7 @@ import { Identifiers, Telemetry } from '../constants';
 import { IDataViewerFactory } from '../data-viewing/types';
 import { NotebookModelChange } from '../interactive-common/interactiveWindowTypes';
 import { KernelSelector } from '../jupyter/kernels/kernelSelector';
-import { NotebookModelEditEvent } from '../notebookStorage/notebookModelEditEvent';
+import { NotebookModelEditEvent } from './notebookModelEditEvent';
 import {
     ICodeCssGenerator,
     IDataScienceErrorHandler,
@@ -61,9 +61,9 @@ import {
     IThemeFinder,
     ITrustService
 } from '../types';
-import { NativeEditor } from './nativeEditor';
+import { NativeEditor } from '../interactive-ipynb/nativeEditor';
 import { getNextUntitledCounter } from './nativeEditorStorage';
-import { NativeEditorSynchronizer } from './nativeEditorSynchronizer';
+import { NativeEditorSynchronizer } from '../interactive-ipynb/nativeEditorSynchronizer';
 import { INotebookStorageProvider } from './notebookStorageProvider';
 
 // Class that is registered as the custom editor provider for notebooks. VS code will call into this class when
