+++ src/test/datascience/jupyter/jupyterCellOutputMimeTypeTracker.unit.test.ts
@@ -12,7 +12,7 @@ import { EventEmitter, Uri } from 'vscode';
 import { getNamesAndValues } from '../../../client/common/utils/enum';
 import { Telemetry } from '../../../client/datascience/constants';
 import { NativeEditor } from '../../../client/datascience/interactive-ipynb/nativeEditor';
-import { NativeEditorProvider } from '../../../client/datascience/interactive-ipynb/nativeEditorProvider';
+import { NativeEditorProvider } from '../../../client/datascience/notebookStorage/nativeEditorProvider';
 import { CellOutputMimeTypeTracker } from '../../../client/datascience/jupyter/jupyterCellOutputMimeTypeTracker';
 import { CellState, ICell, INotebookEditor, INotebookModel } from '../../../client/datascience/types';
