+++ src/test/datascience/nativeEditorViewTracker.unit.test.ts
@@ -7,7 +7,7 @@ import { anything, instance, mock, when } from 'ts-mockito';
 import { EventEmitter, Memento, Uri } from 'vscode';
 import { NotebookModelChange } from '../../client/datascience/interactive-common/interactiveWindowTypes';
 import { NativeEditor } from '../../client/datascience/interactive-ipynb/nativeEditor';
-import { NativeEditorProvider } from '../../client/datascience/interactive-ipynb/nativeEditorProvider';
+import { NativeEditorProvider } from '../../client/datascience/notebookStorage/nativeEditorProvider';
 import { NativeEditorViewTracker } from '../../client/datascience/interactive-ipynb/nativeEditorViewTracker';
 import { NativeEditorNotebookModel } from '../../client/datascience/notebookStorage/notebookModel';
 import { INotebookEditor, INotebookEditorProvider, INotebookModel } from '../../client/datascience/types';
