+++ src/test/datascience/notebook/contentProvider.unit.test.ts
@@ -12,9 +12,9 @@ import type { NotebookContentProvider as VSCodeNotebookContentProvider } from 'v
 import { NotebookCellData } from '../../../../typings/vscode-proposed';
 import { ICommandManager } from '../../../client/common/application/types';
 import { MARKDOWN_LANGUAGE, PYTHON_LANGUAGE } from '../../../client/common/constants';
-import { INotebookStorageProvider } from '../../../client/datascience/interactive-ipynb/notebookStorageProvider';
 import { NotebookContentProvider } from '../../../client/datascience/notebook/contentProvider';
 import { NotebookEditorCompatibilitySupport } from '../../../client/datascience/notebook/notebookEditorCompatibilitySupport';
+import { INotebookStorageProvider } from '../../../client/datascience/notebookStorage/notebookStorageProvider';
 import { CellState, INotebookModel } from '../../../client/datascience/types';
 // tslint:disable: no-any
 suite('DataScience - NativeNotebook ContentProvider', () => {
